Created the Hammer's script
Created the script for the hammer. Updated the TurretCreator, Enemy and GameManager scripts so the cost value for towers and enemies will be 'ints' instead of 'floats'. Moved the csot variable from the turretCreator script to the tower script

+++ Assets/Scripts/Enemy.cs
@@ -11,7 +11,7 @@ public class Enemy : MonoBehaviour
     public Transform end;
 
     public float health;
-    public float value = 3;
+    public int value = 3;
     public float movementSpeed;
 
     public UnityEvent onDeath;
